ticket:1413 changes to compile rasdaman with DENABLE_STRICT=ON

+++ rascontrol_x/src/rascontrolconfig.cc
@@ -265,7 +265,7 @@ std::string RasControlConfig::getCommand() const
     return this->command;
 }
 
-boost::uint32_t RasControlConfig::getRasMgrPort() const
+boost::uint16_t RasControlConfig::getRasMgrPort() const
 {
     return this->rasMgrPort;
 }
