+++ rascontrol_x/src/rascontrolconfig.hh
@@ -86,7 +86,7 @@ public:
 
     std::string getCommand() const;
 
-    boost::uint32_t getRasMgrPort() const;
+    boost::uint16_t getRasMgrPort() const;
 
     void displayHelp() const;
 
@@ -105,7 +105,7 @@ private:
 
     std::string rasMgrHost;/*!< Name of the rasmgr host to which we want to connect.*/
     std::string logConfigFile;/*!< File used to configure logging*/
-    boost::int32_t rasMgrPort;/*!< Port of the rasmgr instance to which we want to connect.*/
+    boost::uint16_t rasMgrPort;/*!< Port of the rasmgr instance to which we want to connect.*/
     boost::int32_t promptMode; /*!< Used to configure the prompt displayed to the user. */
     LoginMode loginMode;
     WorkMode workMode;
